requirements.txt
asyncio
base64
beautifulsoup4
cohere
comtypes
python-dotenv
eel
google-generativeai
google-generativeai-language
groq
keyboard
mtranslate
Pillow
playsound
pyautogui
pycaw
pywhatkit
requests
rich
threading
webbrowser
